requirements.txt
alpaca-trade-api>=3.0.0
PyYAML>=6.0
vaderSentiment>=3.3.2
pytz>=2024.1
